pyproject.toml
[project]
name = "arcagent"
version = "0.1.0"
description = "Custom AI agent framework powered by Claude Code Max tokens"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.50",
    "discord.py>=2.3",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]
